requirements.txt
python-telegram-bot
opencv-python
tensorflow
numpy
python-dotenv
